Day-50/task/4. Performing Read Operations/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
curl_cffi==0.11.4
cycler==0.12.1
Django==5.2.3
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.4
forex-python==1.9.2
frozendict==2.4.6
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Markdown==3.8.1
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.5.0
multitasking==0.0.11
mysql-connector-python==9.3.0
nltk==3.8.1
numpy==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plyer==2.1.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pycoingecko==3.2.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytubefix==9.2.0
pytz==2025.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
speedtest-cli==2.1.3
SQLAlchemy==2.0.41
sqlparse==0.5.3
tqdm==4.67.1
twilio==9.6.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
yfinance==0.2.63
